pyproject.toml
[tool.poetry]
name = "learning-ml"
version = "0.1.0"
description = ""
authors = ["cteng585 <cteng585@stanford.edu>"]
readme = "README.md"
packages = [{include = "learning_ml"}]

[tool.poetry.dependencies]
python = "^3.10"
sklearn = "^0.0"
spacy = "^3.4.2"
tika = "^1.24"
charset-normalizer = "^3.0.0"
python-docx = "^0.8.11"
selenium = "^4.6.0"
typer = "^0.7.0"
parsel = "^1.7.0"
matplotlib = "^3.6.2"
pandas = "^1.5.1"
en-core-web-lg = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.4.1/en_core_web_lg-3.4.1-py3-none-any.whl"}


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
ipykernel = "^6.16.0"
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
